+++ terraform/dev/resources.tf
@@ -1,5 +1,11 @@
 # Hashicorp Providers
-resource "random_string" "this" {
+resource "random_string" "root" {
+  length  = 4
+  special = false
+  upper   = false
+}
+
+resource "random_string" "ksp" {
   length  = 4
   special = false
   upper   = false