+++ terraform/dev/vpc.tf
@@ -6,7 +6,7 @@ module "vpc" {
   # general
   cidr       = "10.0.0.0/16"
   create_igw = true
-  name       = join(var.delimiter, [var.name, var.stage, random_string.this.id])
+  name       = join(var.delimiter, [var.name, var.stage, random_string.root.id])
 
   # flow logs
   enable_flow_log                      = true
@@ -20,6 +20,6 @@ module "vpc" {
 
   tags = merge(
     var.tags,
-    { Name = join(var.delimiter, [var.name, var.stage, random_string.this.id]) }
+    { Name = join(var.delimiter, [var.name, var.stage, random_string.root.id]) }
   )
 }